+++ test/lib/builder/builder.js
@@ -15,6 +15,8 @@ const applicationAPath = path.join(__dirname, "..", "..", "fixtures", "applicati
 const applicationGPath = path.join(__dirname, "..", "..", "fixtures", "application.g");
 const applicationHPath = path.join(__dirname, "..", "..", "fixtures", "application.h");
 const applicationIPath = path.join(__dirname, "..", "..", "fixtures", "application.i");
+const applicationJPath = path.join(__dirname, "..", "..", "fixtures", "application.j");
+const collectionPath = path.join(__dirname, "..", "..", "fixtures", "collection");
 const libraryDPath = path.join(__dirname, "..", "..", "fixtures", "library.d");
 const libraryEPath = path.join(__dirname, "..", "..", "fixtures", "library.e");
 const libraryHPath = path.join(__dirname, "..", "..", "fixtures", "library.h");
@@ -97,7 +99,6 @@ test("Build application.a", (t) => {
 	});
 });
 
-
 test("Build application.a with error", async (t) => {
 	const destPath = "./test/tmp/build/application.a/dest";
 
@@ -106,6 +107,112 @@ test("Build application.a with error", async (t) => {
 		destPath
 	}));
 	t.deepEqual(error.message, `Unknown type 'non existent'`);
+});
+
+test("Build application.a with dependencies", (t) => {
+	const destPath = "./test/tmp/build/application.a/dest-deps";
+	const expectedPath = path.join("test", "expected", "build", "application.a", "dest-deps");
+
+	return builder.build({
+		tree: applicationATree,
+		destPath,
+		excludedTasks: ["generateComponentPreload", "generateStandaloneAppBundle", "generateVersionInfo", "generateLibraryPreload", "escapeNonAsciiCharacters", "generateLibraryManifest"],
+		buildDependencies: true
+	}).then(() => {
+		return findFiles(expectedPath);
+	}).then((expectedFiles) => {
+		// Check for all directories and files
+		assert.directoryDeepEqual(destPath, expectedPath);
+		// Check for all file contents
+		return checkFileContentsIgnoreLineFeeds(expectedFiles, expectedPath, destPath);
+	}).then(() => {
+		t.pass();
+	});
+});
+
+test("Build application.a with dependencies include", (t) => {
+	const destPath = "./test/tmp/build/application.a/dest-deps-incl";
+	const expectedPath = path.join("test", "expected", "build", "application.a", "dest-deps");
+
+	return builder.build({
+		tree: applicationATree,
+		destPath,
+		excludedTasks: ["generateComponentPreload", "generateStandaloneAppBundle", "generateVersionInfo", "generateLibraryPreload", "escapeNonAsciiCharacters", "generateLibraryManifest"],
+		buildDependencies: true, includedDependencies: ["*"]
+	}).then(() => {
+		return findFiles(expectedPath);
+	}).then((expectedFiles) => {
+		// Check for all directories and files
+		assert.directoryDeepEqual(destPath, expectedPath);
+		// Check for all file contents
+		return checkFileContentsIgnoreLineFeeds(expectedFiles, expectedPath, destPath);
+	}).then(() => {
+		t.pass();
+	});
+});
+
+test("Build application.a with dependencies exclude", (t) => {
+	const destPath = "./test/tmp/build/application.a/dest-deps-excl";
+	const expectedPath = path.join("test", "expected", "build", "application.a", "dest-deps-excl");
+
+	return builder.build({
+		tree: applicationATree,
+		destPath,
+		excludedTasks: ["generateComponentPreload", "generateStandaloneAppBundle", "generateVersionInfo", "generateLibraryPreload", "escapeNonAsciiCharacters", "generateLibraryManifest"],
+		buildDependencies: true, excludedDependencies: ["library.d"]
+	}).then(() => {
+		return findFiles(expectedPath);
+	}).then((expectedFiles) => {
+		// Check for all directories and files
+		assert.directoryDeepEqual(destPath, expectedPath);
+		// Check for all file contents
+		return checkFileContentsIgnoreLineFeeds(expectedFiles, expectedPath, destPath);
+	}).then(() => {
+		t.pass();
+	});
+});
+
+test("Build application.a self-contained", (t) => {
+	const destPath = "./test/tmp/build/application.a/dest-self";
+	const expectedPath = path.join("test", "expected", "build", "application.a", "dest-self");
+
+	return builder.build({
+		tree: applicationATree,
+		destPath,
+		excludedTasks: ["generateComponentPreload", "generateVersionInfo"],
+		selfContained: true
+	}).then(() => {
+		return findFiles(expectedPath);
+	}).then((expectedFiles) => {
+		// Check for all directories and files
+		assert.directoryDeepEqual(destPath, expectedPath);
+		// Check for all file contents
+		return checkFileContentsIgnoreLineFeeds(expectedFiles, expectedPath, destPath);
+	}).then(() => {
+		t.pass();
+	});
+});
+
+test("Build application.a with dependencies self-contained", (t) => {
+	const destPath = "./test/tmp/build/application.a/dest-depself";
+	const expectedPath = path.join("test", "expected", "build", "application.a", "dest-depself");
+
+	return builder.build({
+		tree: applicationATree,
+		destPath,
+		excludedTasks: ["generateComponentPreload", "generateVersionInfo", "escapeNonAsciiCharacters", "generateLibraryManifest"],
+		buildDependencies: true,
+		selfContained: true
+	}).then(() => {
+		return findFiles(expectedPath);
+	}).then((expectedFiles) => {
+		// Check for all directories and files
+		assert.directoryDeepEqual(destPath, expectedPath);
+		// Check for all file contents
+		return checkFileContentsIgnoreLineFeeds(expectedFiles, expectedPath, destPath);
+	}).then(() => {
+		t.pass();
+	});
 });
 
 test("Build application.a [dev mode]", (t) => {
@@ -259,6 +366,26 @@ test("Build application.i", (t) => {
 	});
 });
 
+test("Build application.j", (t) => {
+	const destPath = "./test/tmp/build/application.j/dest";
+	const expectedPath = path.join("test", "expected", "build", "application.j", "dest");
+
+	return builder.build({
+		tree: applicationJTree,
+		destPath,
+		excludedTasks: ["createDebugFiles", "generateStandaloneAppBundle", "generateVersionInfo"]
+	}).then(() => {
+		return findFiles(expectedPath);
+	}).then((expectedFiles) => {
+		// Check for all directories and files
+		assert.directoryDeepEqual(destPath, expectedPath);
+		// Check for all file contents
+		return checkFileContentsIgnoreLineFeeds(expectedFiles, expectedPath, destPath);
+	}).then(() => {
+		t.pass();
+	});
+});
+
 test("Build library.d with copyright from .library file", (t) => {
 	const destPath = "./test/tmp/build/library.d/dest";
 	const expectedPath = path.join("test", "expected", "build", "library.d", "dest");
@@ -389,9 +516,9 @@ test("Build theme.j even without an library", (t) => {
 
 test.serial("Cleanup", async (t) => {
 	const BuildContext = require("../../../lib/builder/BuildContext");
-	const projectContext = "project context";
+	const projectContext = {isRootProject: sinon.stub()};
 	const createProjectContextStub = sinon.stub(BuildContext.prototype, "createProjectContext").returns(projectContext);
-	const executeCleanupTasksStub = sinon.stub(BuildContext.prototype, "executeCleanupTasks").returns(projectContext);
+	const executeCleanupTasksStub = sinon.stub(BuildContext.prototype, "executeCleanupTasks").resolves();
 	const applicationType = require("../../../lib/types/application/applicationType");
 	const appBuildStub = sinon.stub(applicationType, "build").resolves();
 
@@ -420,10 +547,11 @@ test.serial("Cleanup", async (t) => {
 	t.deepEqual(appBuildStub.callCount, 1, "Build called once");
 	t.deepEqual(createProjectContextStub.callCount, 1, "One project context got created");
 	const createProjectContextParams = createProjectContextStub.getCall(0).args[0];
+	t.truthy(createProjectContextParams.project, "project object provided");
 	t.truthy(createProjectContextParams.resources.workspace, "resources.workspace object provided");
 	t.truthy(createProjectContextParams.resources.dependencies, "resources.dependencies object provided");
-	t.deepEqual(Object.keys(createProjectContextParams), ["resources"],
-		"resource parameter (and no others) provided");
+	t.deepEqual(Object.keys(createProjectContextParams), ["project", "resources"],
+		"resource and project parameters provided");
 	t.deepEqual(executeCleanupTasksStub.callCount, 1, "Cleanup called once");
 
 	// Error case
@@ -441,41 +569,71 @@ test.serial("Cleanup", async (t) => {
 	t.deepEqual(executeCleanupTasksStub.callCount, 2, "Cleanup called twice");
 });
 
-const applicationATree = {
-	"id": "application.a",
+
+const libraryDTree = {
+	"id": "library.d",
 	"version": "1.0.0",
-	"path": applicationAPath,
+	"path": libraryDPath,
 	"dependencies": [
 		{
-			"id": "library.d",
+			"id": "sap.ui.core-evo",
 			"version": "1.0.0",
-			"path": path.join(applicationAPath, "node_modules", "library.d"),
+			"path": libraryCore,
 			"dependencies": [],
 			"_level": 1,
 			"specVersion": "0.1",
 			"type": "library",
 			"metadata": {
-				"name": "library.d",
-				"namespace": "library/d",
+				"name": "sap.ui.core",
+				"namespace": "sap/ui/core",
 				"copyright": "Some fancy copyright"
 			},
 			"resources": {
 				"configuration": {
 					"paths": {
-						"src": "main/src",
-						"test": "main/test"
+						"src": "main/src"
 					}
 				},
 				"pathMappings": {
-					"/resources/": "main/src",
-					"/test-resources/": "main/test"
+					"/resources/": "main/src"
 				}
 			}
+		}
+	],
+	"_level": 0,
+	"_isRoot": true,
+	"specVersion": "0.1",
+	"type": "library",
+	"metadata": {
+		"name": "library.d",
+		"namespace": "library/d",
+		"copyright": "Some fancy copyright"
+	},
+	"resources": {
+		"configuration": {
+			"paths": {
+				"src": "main/src",
+				"test": "main/test"
+			},
+			"propertiesFileSourceEncoding": "ISO-8859-1"
 		},
+		"pathMappings": {
+			"/resources/": "main/src",
+			"/test-resources/": "main/test"
+		}
+	}
+};
+
+const applicationATree = {
+	"id": "application.a",
+	"version": "1.0.0",
+	"path": applicationAPath,
+	"dependencies": [
+		libraryDTree,
 		{
 			"id": "library.a",
 			"version": "1.0.0",
-			"path": path.join(applicationAPath, "node_modules", "collection", "library.a"),
+			"path": path.join(collectionPath, "library.a"),
 			"dependencies": [],
 			"_level": 1,
 			"specVersion": "0.1",
@@ -501,7 +659,7 @@ const applicationATree = {
 		{
 			"id": "library.b",
 			"version": "1.0.0",
-			"path": path.join(applicationAPath, "node_modules", "collection", "library.b"),
+			"path": path.join(collectionPath, "library.b"),
 			"dependencies": [],
 			"_level": 1,
 			"specVersion": "0.1",
@@ -527,7 +685,7 @@ const applicationATree = {
 		{
 			"id": "library.c",
 			"version": "1.0.0",
-			"path": path.join(applicationAPath, "node_modules", "collection", "library.c"),
+			"path": path.join(collectionPath, "library.c"),
 			"dependencies": [],
 			"_level": 1,
 			"specVersion": "0.1",
@@ -552,10 +710,12 @@ const applicationATree = {
 		}
 	],
 	"_level": 0,
+	"_isRoot": true,
 	"specVersion": "0.1",
 	"type": "application",
 	"metadata": {
-		"name": "application.a"
+		"name": "application.a",
+		"namespace": "application/a"
 	},
 	"resources": {
 		"configuration": {
@@ -570,12 +730,12 @@ const applicationATree = {
 	}
 };
 
-
 const applicationATreeBadType = {
 	"id": "application.a",
 	"version": "1.0.0",
 	"path": applicationAPath,
 	"_level": 0,
+	"_isRoot": true,
 	"specVersion": "0.1",
 	"type": "non existent",
 	"metadata": {
@@ -599,6 +759,7 @@ const applicationGTree = {
 	"version": "1.0.0",
 	"path": applicationGPath,
 	"_level": 0,
+	"_isRoot": true,
 	"specVersion": "0.1",
 	"type": "application",
 	"metadata": {
@@ -634,6 +795,7 @@ const applicationGTreeWithExcludes = {
 	"version": "1.0.0",
 	"path": applicationGPath,
 	"_level": 0,
+	"_isRoot": true,
 	"specVersion": "0.1",
 	"type": "application",
 	"metadata": {
@@ -670,6 +832,7 @@ const applicationGTreeComponentPreloadPaths = {
 	"version": "1.0.0",
 	"path": applicationGPath,
 	"_level": 0,
+	"_isRoot": true,
 	"specVersion": "0.1",
 	"type": "application",
 	"metadata": {
@@ -703,6 +866,7 @@ const applicationHTree = {
 	"version": "1.0.0",
 	"path": applicationHPath,
 	"_level": 0,
+	"_isRoot": true,
 	"specVersion": "0.1",
 	"type": "application",
 	"metadata": {
@@ -764,6 +928,7 @@ const applicationITree = {
 	"version": "1.0.0",
 	"path": applicationIPath,
 	"_level": 0,
+	"_isRoot": true,
 	"specVersion": "0.1",
 	"type": "application",
 	"metadata": {
@@ -787,56 +952,32 @@ const applicationITree = {
 	}
 };
 
-const libraryDTree = {
-	"id": "library.d",
+const applicationJTree = {
+	"id": "application.j",
 	"version": "1.0.0",
-	"path": libraryDPath,
-	"dependencies": [
-		{
-			"id": "sap.ui.core-evo",
-			"version": "1.0.0",
-			"path": libraryCore,
-			"dependencies": [],
-			"_level": 1,
-			"specVersion": "0.1",
-			"type": "library",
-			"metadata": {
-				"name": "sap.ui.core",
-				"namespace": "sap/ui/core",
-				"copyright": "Some fancy copyright"
-			},
-			"resources": {
-				"configuration": {
-					"paths": {
-						"src": "main/src"
-					}
-				},
-				"pathMappings": {
-					"/resources/": "main/src"
-				}
-			}
-		}
-	],
+	"path": applicationJPath,
 	"_level": 0,
+	"_isRoot": true,
 	"specVersion": "0.1",
-	"type": "library",
+	"type": "application",
 	"metadata": {
-		"name": "library.d",
-		"namespace": "library/d",
-		"copyright": "Some fancy copyright"
+		"name": "application.j",
+		"namespace": "application/j"
 	},
+	"dependencies": [],
 	"resources": {
 		"configuration": {
 			"paths": {
-				"src": "main/src",
-				"test": "main/test"
+				"webapp": "webapp"
 			},
 			"propertiesFileSourceEncoding": "ISO-8859-1"
 		},
 		"pathMappings": {
-			"/resources/": "main/src",
-			"/test-resources/": "main/test"
+			"/": "webapp"
 		}
+	},
+	"builder": {
+		"bundles": []
 	}
 };
 
@@ -871,6 +1012,7 @@ const libraryETree = {
 		}
 	],
 	"_level": 0,
+	"_isRoot": true,
 	"specVersion": "0.1",
 	"type": "library",
 	"metadata": {
@@ -924,6 +1066,7 @@ const libraryHTree = {
 		}
 	],
 	"_level": 0,
+	"_isRoot": true,
 	"specVersion": "0.1",
 	"type": "library",
 	"metadata": {
@@ -1010,6 +1153,7 @@ const libraryITree = {
 		cloneProjectTree(libraryDTree)
 	],
 	"_level": 0,
+	"_isRoot": true,
 	"specVersion": "0.1",
 	"type": "library",
 	"metadata": {
@@ -1037,6 +1181,7 @@ const libraryJTree = {
 	"path": libraryJPath,
 	"dependencies": [],
 	"_level": 0,
+	"_isRoot": true,
 	"specVersion": "0.1",
 	"type": "library",
 	"metadata": {
@@ -1062,6 +1207,7 @@ const themeJTree = {
 	"path": themeJPath,
 	"dependencies": [],
 	"_level": 0,
+	"_isRoot": true,
 	"specVersion": "0.1",
 	"type": "library",
 	"metadata": {