+++ test/lib/types/AbstractBuilder.js
@@ -26,10 +26,12 @@ const applicationBTree = {
 	dependencies: [],
 	builder: {},
 	_level: 0,
+	_isRoot: true,
 	specVersion: "0.1",
 	type: "application",
 	metadata: {
-		name: "application.b"
+		name: "application.b",
+		namespace: "application/b"
 	},
 	resources: {
 		configuration: {
@@ -73,6 +75,16 @@ class CustomBuilder extends AbstractBuilder {
 		this.addTask("myStandardTask", function() {});
 		this.addTask("createDebugFiles", function() {});
 		this.addTask("replaceVersion", function() {});
+	}
+}
+
+class EmptyBuilder extends AbstractBuilder {
+	constructor({project, resourceCollections}) {
+		super({parentLogger, project, resourceCollections});
+	}
+
+	addStandardTasks({resourceCollections, project}) {
+		// None - like the ModuleBuilder
 	}
 }
 
@@ -160,8 +172,10 @@ test("Instantiation with custom task and unknown beforeTask", (t) => {
 });
 
 test.serial("Instantiation with custom task", (t) => {
-	const customTask = function() {};
-	sinon.stub(taskRepository, "getTask").returns(customTask);
+	sinon.stub(taskRepository, "getTask").returns({
+		task: function() {},
+		specVersion: undefined
+	});
 
 	const project = clone(applicationBTree);
 	project.builder = {
@@ -177,9 +191,55 @@ test.serial("Instantiation with custom task", (t) => {
 		"Order of tasks is correct");
 });
 
+test.serial("Instantiation of empty builder with custom tasks", (t) => {
+	sinon.stub(taskRepository, "getTask").returns({
+		task: function() {},
+		specVersion: undefined
+	});
+
+	const project = clone(applicationBTree);
+	project.builder = {
+		customTasks: [{
+			name: "myTask"
+		}, {
+			name: "myTask2",
+			beforeTask: "myTask"
+		}]
+	};
+	const customBuilder = new EmptyBuilder({project});
+	t.truthy(customBuilder.tasks["myTask"], "Custom task has been added to task array");
+	t.truthy(customBuilder.tasks["myTask2"], "Custom task 2 has been added to task array");
+	t.deepEqual(customBuilder.taskExecutionOrder,
+		["myTask2", "myTask"],
+		"Order of tasks is correct");
+});
+
+test.serial("Instantiation of empty builder with 2nd custom tasks defining neither beforeTask nor afterTask", (t) => {
+	sinon.stub(taskRepository, "getTask").returns({
+		task: function() {},
+		specVersion: "2.0"
+	});
+
+	const project = clone(applicationBTree);
+	project.builder = {
+		customTasks: [{
+			name: "myTask"
+		}, {
+			name: "myTask2" // should define before- or afterTask
+		}]
+	};
+	const error = t.throws(() => {
+		new EmptyBuilder({project});
+	}, Error);
+	t.deepEqual(error.message, `Custom task definition myTask2 of project application.b defines ` +
+		`neither a "beforeTask" nor an "afterTask" parameter. One must be defined.`, "Correct exception thrown");
+});
+
 test.serial("Instantiation with custom task defined three times", (t) => {
-	const customTask = function() {};
-	sinon.stub(taskRepository, "getTask").returns(customTask);
+	sinon.stub(taskRepository, "getTask").returns({
+		task: function() {},
+		specVersion: "2.0"
+	});
 
 	const project = clone(applicationBTree);
 	project.builder = {
@@ -206,11 +266,15 @@ test.serial("Instantiation with custom task defined three times", (t) => {
 test.serial("Instantiation with custom task: Custom task called correctly", (t) => {
 	const customTask = function({workspace, dependencies, options}) {
 		t.deepEqual(options.projectName, "application.b", "Correct project name passed to custom task");
+		t.deepEqual(options.projectNamespace, "application/b", "Correct project namespace passed to custom task");
 		t.deepEqual(options.configuration, "pony", "Correct configuration passed to custom task");
 		t.deepEqual(workspace, "myWorkspace", "Correct workspace passed to custom task");
 		t.deepEqual(dependencies, "myDependencies", "Correct dependency collection passed to custom task");
 	};
-	sinon.stub(taskRepository, "getTask").returns(customTask);
+	sinon.stub(taskRepository, "getTask").returns({
+		task: customTask,
+		specVersion: "2.0"
+	});
 
 	const project = clone(applicationBTree);
 	project.builder = {
@@ -236,8 +300,14 @@ test.serial("Instantiation with custom task: Two custom tasks called correctly",
 		t.deepEqual(options.configuration, "donkey", "Correct configuration passed to second custom task");
 	};
 	const stubGetTask = sinon.stub(taskRepository, "getTask");
-	stubGetTask.onCall(0).returns(customTask1);
-	stubGetTask.onCall(1).returns(customTask2);
+	stubGetTask.onCall(0).returns({
+		task: customTask1,
+		specVersion: "2.0"
+	});
+	stubGetTask.onCall(1).returns({
+		task: customTask2,
+		specVersion: "2.0"
+	});
 
 	const project = clone(applicationBTree);
 	project.builder = {