+++ test/lib/types/application/ApplicationFormatter.js
@@ -17,7 +17,8 @@ const applicationBTree = {
 	path: applicationBPath,
 	dependencies: [],
 	_level: 0,
-	specVersion: "0.1",
+	_isRoot: true,
+	specVersion: "2.0",
 	type: "application",
 	metadata: {
 		name: "application.b"
@@ -102,7 +103,41 @@ test("validate: empty encoding", async (t) => {
 	const applicationFormatter = new ApplicationFormatter({project: myProject});
 
 	await applicationFormatter.validate(myProject);
-	t.deepEqual(myProject.resources.configuration.propertiesFileSourceEncoding, "ISO-8859-1", "default resources encoding is set");
+	t.deepEqual(myProject.resources.configuration.propertiesFileSourceEncoding, "UTF-8",
+		"default resources encoding is set");
+});
+
+test("validate: empty encoding - legacy specVersion 0.1", async (t) => {
+	const myProject = clone(applicationBTree);
+	myProject.specVersion = "0.1";
+	delete myProject.resources.configuration.propertiesFileSourceEncoding;
+	const applicationFormatter = new ApplicationFormatter({project: myProject});
+
+	await applicationFormatter.validate(myProject);
+	t.deepEqual(myProject.resources.configuration.propertiesFileSourceEncoding, "ISO-8859-1",
+		"default resources encoding is set");
+});
+
+test("validate: empty encoding - legacy specVersion 1.0", async (t) => {
+	const myProject = clone(applicationBTree);
+	myProject.specVersion = "1.0";
+	delete myProject.resources.configuration.propertiesFileSourceEncoding;
+	const applicationFormatter = new ApplicationFormatter({project: myProject});
+
+	await applicationFormatter.validate(myProject);
+	t.deepEqual(myProject.resources.configuration.propertiesFileSourceEncoding, "ISO-8859-1",
+		"default resources encoding is set");
+});
+
+test("validate: empty encoding - legacy specVersion 1.1", async (t) => {
+	const myProject = clone(applicationBTree);
+	myProject.specVersion = "1.1";
+	delete myProject.resources.configuration.propertiesFileSourceEncoding;
+	const applicationFormatter = new ApplicationFormatter({project: myProject});
+
+	await applicationFormatter.validate(myProject);
+	t.deepEqual(myProject.resources.configuration.propertiesFileSourceEncoding, "ISO-8859-1",
+		"default resources encoding is set");
 });
 
 test("validate: test invalid encoding", async (t) => {
@@ -111,8 +146,8 @@ test("validate: test invalid encoding", async (t) => {
 	const applicationFormatter = new ApplicationFormatter({project: myProject});
 
 	const error = await t.throwsAsync(applicationFormatter.validate(myProject));
-	t.is(error.message, `Invalid properties file encoding specified for project application.b: encoding provided: test. Must be either "ISO-8859-1" or "UTF-8".`,
-		"Missing source directory caused error");
+	t.is(error.message, `Invalid properties file encoding specified for project application.b. Encoding provided: ` +
+		`test. Must be either "ISO-8859-1" or "UTF-8".`, "Missing source directory caused error");
 });
 
 function createMockProject() {
@@ -130,16 +165,24 @@ function createMockProject() {
 	};
 }
 
+test("getSourceBasePath: posix", async (t) => {
+	const myProject = clone(applicationBTree);
+	myProject.path = "my/pony";
+	const applicationFormatter = new ApplicationFormatter({project: myProject});
+
+	const sourceBasePath = applicationFormatter.getSourceBasePath(true);
+	t.is(sourceBasePath, "my/pony/webapp", "correct path");
+});
+
 test("format: No 'sap.app' configuration found", async (t) => {
 	const project = createMockProject();
 	const applicationFormatter = new ApplicationFormatter({project});
 	sinon.stub(applicationFormatter, "validate").resolves();
 	sinon.stub(applicationFormatter, "getManifest").resolves({content: {}, fsPath: {}});
 
-	await applicationFormatter.format();
-	t.deepEqual(project.resources.pathMappings["/"], "webapp", "path mappings is set");
-	t.falsy(project.metadata.namespace,
-		"namespace is falsy since getManifest resolves with an empty object");
+	const error = await t.throwsAsync(applicationFormatter.format());
+	t.deepEqual(error.message, "No sap.app/id configuration found in manifest.json of project projectName",
+		"Rejected with correct error message");
 });
 
 test("format: No application id in 'sap.app' configuration found", async (t) => {
@@ -148,10 +191,9 @@ test("format: No application id in 'sap.app' configuration found", async (t) => 
 	sinon.stub(applicationFormatter, "validate").resolves();
 	sinon.stub(applicationFormatter, "getManifest").resolves({content: {"sap.app": {}}});
 
-	await applicationFormatter.format();
+	const error = await t.throwsAsync(applicationFormatter.format());
+	t.deepEqual(error.message, "No sap.app/id configuration found in manifest.json of project projectName");
 	t.deepEqual(project.resources.pathMappings["/"], "webapp", "path mappings is set");
-	t.falsy(project.metadata.namespace,
-		"namespace is falsy since getManifest resolves with an empty object");
 });
 
 test("format: set namespace to id", async (t) => {
@@ -220,11 +262,12 @@ test.serial("getManifest: result is cached", async (t) => {
 
 	const ApplicationFormatter = mock.reRequire("../../../../lib/types/application/ApplicationFormatter");
 	const libraryFormatter = new ApplicationFormatter({project: myProject});
-
 	const expectedPath = path.join(applicationBPath, "webapp", "manifest.json");
+
 	const {content, fsPath} = await libraryFormatter.getManifest();
 	t.deepEqual(content, {pony: "no unicorn"}, "Correct result on first call");
 	t.deepEqual(fsPath, expectedPath, "Correct manifest.json path returned on first call");
+
 	const {content: content2, fsPath: fsPath2} = await libraryFormatter.getManifest();
 	t.deepEqual(content2, {pony: "no unicorn"}, "Correct result on second call");
 	t.deepEqual(fsPath2, expectedPath, "Correct manifest.json path returned on second call");
@@ -239,7 +282,8 @@ const applicationHTree = {
 	path: applicationHPath,
 	dependencies: [],
 	_level: 0,
-	specVersion: "0.1",
+	_isRoot: true,
+	specVersion: "2.0",
 	type: "application",
 	metadata: {
 		name: "application.h"
@@ -278,7 +322,7 @@ test("namespace: detect namespace from pom.xml via ${appId} from properties", as
 	myProject.resources.configuration.paths.webapp = "webapp-properties.appId";
 	const applicationFormatter = new ApplicationFormatter({project: myProject});
 
-	await applicationFormatter.format();
-	t.falsy(myProject.metadata.namespace,
-		"namespace is falsy since getManifest resolves with an empty object");
+	const error = await t.throwsAsync(applicationFormatter.format());
+	t.deepEqual(error.message, "Failed to resolve namespace of project application.h: \"${appId}\"" +
+		" couldn't be resolved from maven property \"appId\" of pom.xml of project application.h");
 });